complicated fix to handle heterogeneous WIM sites
couple of things added.  First, expect to get lanes count from db
neighbor query.  Use the lanes count to figure out which sites to
keep.  the idea is to keep sites with about the same number of lanes
that I am processing.

If there are only a fewer number of lanes, keep the max of those

If there are more lanes, at the moment I keep them all down to the
number of lanes at the site, but I might fix that after
testing...we'll see.

Second, when merging in sites, only keep those with a proper set of
data.  That is, I expect to have truck imputations and also the
aggregate values from the speed db table.  I've found that some sites
have broken aggregate values reports.  Those sites get dropped from
this imputation step.

This also will probably need to be revisited, if there is a case in
which there is just one site available and it doesn't have a proper
set of data, I'd probably want to keep it.