make two streams?
got another crash.  staring, it is definitely a node error, and it is
definitely a problem writing to a closed stream.

so make an err stream and a log stream, both pointing at the same file,
and assign each to stdout and stderr as appropriate.

I think as stdout dies, it closes the stream, and then stderr tries to
write, but the stream is already closed??