look for imputed file in the right directory
There was a problem looking for district 5 files, because caltrans
extended the vdsids to more than the usual 6 digits, so I had dupes, so
the dir() command was finding irrelevant imputed files.

this fix makes the path variable the same as the leaf directory where
the original file is found, rather than the root directory for the
district.  This way a recursive search will not descend into similar
named locations.

Will still fail if, say, you have two detectors with similar numbers in
the same subdirectory, but that hasn't happened in practice yet.